Collected Auth and Firebase in layout. Clean up unused stuff
