+++ src/routes/[pin]/[round]/+page.server.js
@@ -1,10 +1,10 @@
-import { admin, validateUserAndGetUid } from "$lib/admin.server"
+import { admin } from "$lib/admin.server"
 
 /** @type {import('@sveltejs/kit').Actions} */
 export const actions = {
   submit: async (event) => {
     const formData = await event.request.formData()
-    const uid = await validateUserAndGetUid(formData.get("user-id-token"))
+    const uid = event.cookies.get("firebaseuid")
     const auctionRef = admin.database().ref("auctions/" + event.params.pin)
     await auctionRef.child("/bids/" + uid).set(formData.get("bids"))
     await auctionRef.child("/readys/" + uid).set(event.params.round)
