+++ src/routes/api/cookie/+server.js
@@ -1,17 +1,15 @@
 import { validateUserAndGetUid } from "$lib/admin.server"
 import { redirect } from "@sveltejs/kit"
 
-/** @type {import('./$types').RequestHandler} */
-export async function POST({ cookies, request }) {
-  const { userIdToken } = await request.json()
-  const uid = await validateUserAndGetUid(userIdToken)
-  cookies.set("uid", uid, {
+/** @type {import('@sveltejs/kit').RequestHandler} */
+export async function POST(event) {
+  const requestBody = await event.request.json()
+  const uid = await validateUserAndGetUid(requestBody.useridtoken)
+  event.cookies.set("firebaseuid", uid, {
     path: "/",
     maxAge: 60 * 60 * 24 * 2, // say, 2 days
   })
-  console.log("throw redirect(301, `/cookied`)")
   throw redirect(302, `/cookied`)
-  // return new Response(uid, { status: 201 })
 }
 
 /** @type {import('./$types').RequestHandler} */
